feat(document_review): add auto checker and tests
to save time and resource, this script will pull back all documentation that is currently out of date. You can automate the process of reviewing by i) reviewing the document PROPERLY, ii) running this script and either fixing the date or amending manually. This is a work in progress and just appears as an outline for the expected functionality.