⚡️  Add retry logic to HTTP POST requests
This commit adds retry logic to the `__http_post` method in the `ReportsService` class. The `Retry` object from the `requests` library is used to automatically retry failed requests up to 3 times, with a backoff factor of 1 second between retries.

The `status_forcelist` parameter is set to `[500, 502, 503, 504]` to retry requests that fail due to server errors. The `method_whitelist` parameter is set to `["POST"]` to retry only POST requests. This improves the reliability of the `ReportsService` class when making HTTP POST requests to the API.