:recycle: Update string formatting in logging
This commit changes the f-string based logging messages to old style %-formatting in the `Auth0Service` class. The decision to shift towards older style is based on its advantage when it comes to delayed string interpolation, which is preferable for potentially expensive operations like logging. This change improves performance especially when those log levels are not enabled.