:recycle: Refactor test methods to use inline mock data
This commit modifies the `test_get_usernames_from_csv_ignoring_bots_only_bots` and `test_get_usernames_from_csv_ignoring_bots_and_collaborators` in `TestDormantGitHubUsers`. It replaces the usage of `@patch` decorator before the function with an inline mock object using Python's context manager. The reason for this change is to increase readability and reduce the complexity of the code.