:sparkles: Add Slack service to identify_dormant_github_users script
The commit includes various updates to the `identify_dormant_github_users.py` script. The changes allow it to leverage the SlackService for notifying dormant GitHub users. It does this by sending a message to a specified Slack channel with the list of identified dormant users from GitHub, improving communication effectiveness.

Key changes include:
1. Added import statement for SlackService.
2. Updated setup_services function to also return an instance of SlackService initialized with ADMIN_SLACK_TOKEN.
3. Implemented a new function, message_to_slack_channel, which creates a Slack message with the number and list of dormant users.
4. Updated main code block to integrate Slack messaging into the dormant user identification process.

A new environment variable, ADMIN_SLACK_TOKEN, is required for initializing the SlackService. An exception will be raised if it's not set.