:sparkles: Add IAM resources for GitHub Dormant User
This commit introduces several new AWS IAM and S3 resources to manage a 'dormant' GitHub user. It includes creation of an IAM access key associated with the user, a bucket policy allowing the user to perform certain actions on an S3 bucket, and GitHub Actions secrets that store the IAM user's access keys. The intention behind these changes is to establish an infrastructure for handling inactive or 'dormant' GitHub users.