:fire: Remove AWS keys from GitHub secrets
This commit removes the resources related to `github_actions_secret` for `AWS_ACCESS_KEY` and `AWS_SECRET_KEY`. The decision to omit these was likely based on a change in the way access control is handled, possibly due to security considerations. As a result of these changes, the AWS keys will no longer be available as GitHub secrets in the "operations-engineering" repository. Further updates or adjustments may be necessary depending on how the AWS IAM Access Keys are used within the application.