:sparkles: Add manual GitHub Actions workflow for identifying dormant users
This commit introduces a new experimental workflow in GitHub Actions that can manually be triggered to identify dormant GitHub users. It's designed to run on Ubuntu latest with Python 3.12 and installs dependencies using pipenv. The execution of the main script uses environment variables to authenticate necessary services.

The workflow, once stable, will be integrated into the primary process of managing dormant user accounts.