:hammer_and_wrench: Refactor dormant user identification logic
This commit refactors the process of identifying dormant GitHub users by moving the dormancy calculation into the DormantGitHubUser class. The calculation is now performed using a new `is_dormant` property, which checks the last activity timestamp against a threshold. It also includes removal of unused methods and improves the presentation of results in stdout by differentiating between deleted and not deleted users. The logic for fetching the last Auth0 activity from CloudWatch Logs has been expanded to execute a query and then fetch and parse the results.