:recycle: Refactor DormantUserEnvironment class and extend tests
This commit refactors the `DormantUserEnvironment` class to use an initializer for setting environmental variables. We now raise an exception if any of these environment variables (`GH_ADMIN_TOKEN`, `AUTH0_CLIENT_SECRET`, `AUTH0_CLIENT_ID`, or `ADMIN_SLACK_TOKEN`) are not set or empty.

We have also expanded our test coverage to include cases for `DormantUserEnvironment`. These added tests check the correct initialization of environment variables and the handling of cases where one or more of these variables are missing or empty.

A knock-on effect of this change meant that we had to remove some unused imports and adjust various test case definitions in `test_identify_dormant_github_users.py`.