:wrench: Add new GitHub Team terraform pipeline and support team
This commit contains the ability to manage GitHub teams in code. I've decided to create a support team to translate to more efficient operations, better security practices, and improved ability to manage access to sensitive systems and data. It ensures that as support staff changes, their access to necessary resources is managed promptly and accurately, which is crucial for both operational efficiency and security in a support role.