:sparkles: Add CloudWatchService with insights query functionality
This commit adds a new class, `CloudWatchService`, to the services. This class simplifies interaction with AWS CloudWatch Logs: it initiates connection to AWS, provides a method for running Insights Queries and polling for results, and also defines a specific method for getting all Auth0 users that appear in tenants within a specified timeframe. It's anticipated this refactoring will improve readability and maintainability of code dealing with CloudWatch Logs.