:bug: Refactor CloudWatchService timeout condition
Refactored the timeout condition in the CloudWatchService class to improve flow control. The previous 'elif' has been replaced with an 'if' statement to make sure the timeout check is performed regardless of the query status. This could prevent potential infinite loops if a complete response is not received for any reason.