:sparkles: Add a Flask template repository to increase velocity
The idea behind this is to give the Operations Engineering team the ability to quickly create applications based off a tried and trusted flask template. If you look at all of our templates, they all appear identical in a number of areas. By creating a boilerplate template, we should be improving velocity and standardising best practice.